[PDI-15328] PDI needs to error out and log an error when a plugin is missing at run time - fixed and moved integration tests - fixed checkstyle violations
