+++ engine/src/org/pentaho/di/job/entries/missing/messages/messages_en_US.properties
@@ -1,2 +1,2 @@
 
-MissingEntry.Log.CannotRunJob=Can\'t run job due to plugin missing
+MissingEntry.Log.CannotRunJob=Can''t run job due to plugin missing
