+++ engine/src/org/pentaho/di/trans/steps/missing/MissingTrans.java
@@ -54,9 +54,9 @@ public class MissingTrans extends DummyTransMeta {
   public void setMissingPluginId( String missingPluginId ) {
     this.missingPluginId = missingPluginId;
   }
-  
+
   public StepInterface getStep( StepMeta stepMeta, StepDataInterface stepDataInterface, int cnr, TransMeta tr,
       Trans trans ) {
-      return new MissingTransStep( stepMeta, stepDataInterface, cnr, tr, trans );
-    }
+    return new MissingTransStep( stepMeta, stepDataInterface, cnr, tr, trans );
+  }
 }
