+++ engine/src/org/pentaho/di/trans/steps/missing/messages/messages_en_US.properties
@@ -3,6 +3,6 @@
 ##  MissingTransStep
 ##
 #####################################################################
-MissingTransStep.Log.CannotRunTrans=Can\'t run transformation due to plugin missing
+MissingTransStep.Log.CannotRunTrans=Can''t run transformation due to plugin missing
